# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /broker/connect:
    post:
      operationId: connectBroker
      tags: [broker]
      summary: Provision and connect a live MetaApi account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrokerConnectInput"
      responses:
        "200":
          description: Connected account details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrokerAccount"
        "400":
          $ref: "#/components/responses/BadRequest"
        "502":
          $ref: "#/components/responses/UpstreamError"
  /broker/balance:
    get:
      operationId: getBrokerBalance
      tags: [broker]
      summary: Read live account information and open positions
      parameters:
        - $ref: "#/components/parameters/AccountId"
      responses:
        "200":
          description: Live broker snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrokerAccount"
        "400":
          $ref: "#/components/responses/BadRequest"
        "502":
          $ref: "#/components/responses/UpstreamError"
  /broker/positions:
    get:
      operationId: getBrokerPositions
      tags: [broker]
      summary: Read live broker positions
      parameters:
        - $ref: "#/components/parameters/AccountId"
      responses:
        "200":
          description: Live broker positions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Position"
        "400":
          $ref: "#/components/responses/BadRequest"
        "502":
          $ref: "#/components/responses/UpstreamError"
  /broker/market-data:
    get:
      operationId: getBrokerMarketData
      tags: [broker]
      summary: Read live quote and broker symbol health
      parameters:
        - $ref: "#/components/parameters/AccountId"
        - name: symbol
          in: query
          required: true
          schema:
            type: string
            minLength: 1
        - name: realSymbol
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Live market-data health
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketDataHealth"
        "400":
          $ref: "#/components/responses/BadRequest"
        "502":
          $ref: "#/components/responses/UpstreamError"
  /broker/close-all:
    post:
      operationId: closeAllPositions
      tags: [broker]
      summary: Close every open position for an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountActionInput"
      responses:
        "200":
          description: Close operation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "502":
          $ref: "#/components/responses/UpstreamError"
  /broker/disconnect:
    post:
      operationId: disconnectBroker
      tags: [broker]
      summary: Undeploy a MetaApi broker account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountActionInput"
      responses:
        "200":
          description: Disconnection result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisconnectResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "502":
          $ref: "#/components/responses/UpstreamError"
...
components:
  parameters:
    AccountId:
      name: accountId
      in: query
      required: true
      schema:
        type: string
        minLength: 1
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    UpstreamError:
      description: Upstream service failure
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: [string, "null"]
      required: [error]
    BrokerConnectInput:
      type: object
      properties:
        login:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
        server:
          type: string
          minLength: 1
        brokerName:
          type: string
          minLength: 1
      required: [login, password, server, brokerName]
    BrokerAccount:
      type: object
      properties:
        accountId:
          type: string
        brokerName:
          type: [string, "null"]
        server:
          type: [string, "null"]
        accountType:
          type: [string, "null"]
        connectionState:
          type: [string, "null"]
        dataSource:
          type: [string, "null"]
        balance:
          type: [number, "null"]
        equity:
          type: [number, "null"]
        leverage:
          type: [number, "null"]
        leverageDisplay:
          type: [string, "null"]
        currency:
          type: [string, "null"]
        margin:
          type: [number, "null"]
        freeMargin:
          type: [number, "null"]
        marginLevel:
          type: [number, "null"]
        connected:
          type: boolean
        positions:
          type: array
          items:
            $ref: "#/components/schemas/Position"
      required: [accountId, connected, positions]
    MarketDataHealth:
      type: object
      properties:
        symbol:
          type: string
        realSymbol:
          type: string
        bid:
          type: [number, "null"]
        ask:
          type: [number, "null"]
        spread:
          type: [number, "null"]
        spreadPoints:
          type: [number, "null"]
        lastTick:
          type: [string, "null"]
        lastCandle:
          type: [string, "null"]
        dataSource:
          type: string
        dataAgeSeconds:
          type: [number, "null"]
        status:
          type: string
          enum: [LIVE, STALE, ERROR]
        marketStatus:
          type: [string, "null"]
        specification:
          type: object
          additionalProperties: true
      required: [symbol, realSymbol, dataSource, status]
    Position:
      type: object
      properties:
        id:
          type: [string, "null"]
        symbol:
          type: [string, "null"]
        type:
          type: [string, "null"]
        volume:
          type: [number, "null"]
        openPrice:
          type: [number, "null"]
        currentPrice:
          type: [number, "null"]
        stopLoss:
          type: [number, "null"]
        takeProfit:
          type: [number, "null"]
        profit:
          type: [number, "null"]
        swap:
          type: [number, "null"]
        time:
          type: [string, "null"]
      required: [symbol, type, volume, openPrice]
    AccountActionInput:
      type: object
      properties:
        accountId:
          type: string
          minLength: 1
      required: [accountId]
    ActionResult:
      type: object
      properties:
        ok:
          type: boolean
        message:
          type: string
        count:
          type: number
        failed:
          type: array
          items:
            type: object
            properties:
              positionId:
                type: string
              ok:
                type: boolean
              error:
                type: [string, "null"]
            required: [positionId, ok, error]
        remaining:
          type: array
          items:
            type: string
        halted:
          type: boolean
      required: [ok, message, count, failed, remaining, halted]
    DisconnectResult:
      type: object
      properties:
        ok:
          type: boolean
        message:
          type: string
        count:
          type: number
        accountId:
          type: string
        state:
          type: [string, "null"]
        connectionStatus:
          type: [string, "null"]
        connected:
          type: boolean
      required: [ok, message, count, accountId, connected]
...